1 page version of the program
wrote this very hungover today. didn't actually follow the principals of TDD to create it (or rather i didn't use rspec/do unit tests, just a lot of of testing in terminal), which was the original point of it, so that was pretty stupid of me, but it works and i think it's robust.